Create `RouteAggregator.tsx`
