+++ src/app/outgoing/traefik/components/TrafiServiceCard.tsx
@@ -7,13 +7,14 @@ import {DEMO_MODE} from "@/app/outgoing/traefik/config";
 import {applyDemoDomainOverride} from "@/app/utils";
 
 export const TrafiServiceCard: FC<TrafiServiceCardProps> = ({trafiService}: TrafiServiceCardProps) => {
-    const serviceName = trafiService.name
+    let serviceName = trafiService.name
     const serviceRoute = trafiService.rule ? trafiService.rule : "no-op"
     let serviceDomain = serviceRoute.split("://")[1]
     const thumbnailUrl = trafiService.thumbnailUrl
     const faviconUrl = serviceRoute + '/favicon.ico'
     if (DEMO_MODE) {
         serviceDomain = applyDemoDomainOverride(serviceDomain);
+        serviceName = applyDemoDomainOverride(serviceName);
     }
     return (
             <Card
